completed my homepage component
